+++ lib/src/serializer/repo.dart
@@ -26,12 +26,14 @@ class SerializerRepo {
   final Map<Type, Serializer> _mapperType = {};
   final Map<String, Serializer> _mapperString = {};
 
-  SerializerRepo({this.typeInfoKey: defaultTypeInfoKey});
+  SerializerRepo({String typeKey: defaultTypeInfoKey}) : _typeKey = typeKey;
 
   /**
    * Key added to a map when serializing an Object
    */
-  final String typeInfoKey;
+  final String _typeKey;
+
+  String getTypeKey() => _typeKey;
 
   /**
    * Return a [Serializer] for a Type
@@ -76,35 +78,30 @@ class SerializerRepo {
   /**
    * Convert the given [Object] to a serialized [Object], [Map] or [List]
    *
-   * If [withTypeInfo] is set to true, the serialized [Object] will contain a key ([typeInfoKey])
+   * If [withType] is set to true, the serialized [Object] will contain a key ([typeKey])
    * with the type of the object as a value.
    *
-   * The [typeInfoKey] can be override using the [useTypeInfoKey] option.
+   * The [typeKey] can be override using the [typeKey] option.
    */
-  dynamic serialize(dynamic object,
-      {bool withTypeInfo: false, String useTypeInfoKey}) {
-    useTypeInfoKey ??= typeInfoKey;
+  dynamic serialize(dynamic object, {bool withType: false, String typeKey}) {
+    typeKey ??= _typeKey;
     if (object is Iterable) {
       return encode(object
           .map((obj) => _to(obj,
-              type: obj.runtimeType,
-              withTypeInfo: withTypeInfo,
-              useTypeInfoKey: useTypeInfoKey))
+              type: obj.runtimeType, withType: withType, typeKey: typeKey))
           .toList());
     } else if (object is Map) {
       Map map = {};
       for (var key in object.keys) {
         map[key] = _to(object[key],
             type: object[key].runtimeType,
-            withTypeInfo: withTypeInfo,
-            useTypeInfoKey: useTypeInfoKey);
+            withType: withType,
+            typeKey: typeKey);
       }
       return encode(map);
     }
     return encode(_to(object,
-        type: object.runtimeType,
-        withTypeInfo: withTypeInfo,
-        useTypeInfoKey: useTypeInfoKey));
+        type: object.runtimeType, withType: withType, typeKey: typeKey));
   }
 
   /**
@@ -112,20 +109,20 @@ class SerializerRepo {
    *
    * If the [type] option is specified, it will be used to get the correct [Serializer].
    *
-   * If not, it will look at if the object contain the [typeInfoKey] and will use it to get the correct [Serializer].
+   * If not, it will look at if the object contain the [typeKey] and will use it to get the correct [Serializer].
    *
-   * The [typeInfoKey] can be override using the [useTypeInfoKey] option.
+   * The [typeKey] can be override using the [typeKey] option.
    */
-  dynamic deserialize(dynamic object, {Type type, String useTypeInfoKey}) {
-    useTypeInfoKey ??= typeInfoKey;
+  dynamic deserialize(dynamic object, {Type type, String typeKey}) {
+    typeKey ??= _typeKey;
     final decoded = object is String ? decode(object) : object;
 
     if (decoded is Iterable) {
       return decoded
-          .map((obj) => _from(obj, type: type, useTypeInfoKey: useTypeInfoKey))
+          .map((obj) => _from(obj, type: type, typeKey: typeKey))
           .toList();
     }
-    return _from(decoded, type: type, useTypeInfoKey: useTypeInfoKey);
+    return _from(decoded, type: type, typeKey: typeKey);
   }
 
   ///@nodoc
@@ -135,8 +132,8 @@ class SerializerRepo {
   dynamic decode(dynamic object) => object;
 
   dynamic _to(dynamic object,
-      {Type type, bool withTypeInfo: false, String useTypeInfoKey}) {
-    useTypeInfoKey ??= typeInfoKey;
+      {Type type, bool withType: false, String typeKey}) {
+    typeKey ??= _typeKey;
     if (object is String || object is num) {
       return object;
     }
@@ -146,17 +143,16 @@ class SerializerRepo {
       throw new Exception("Cannot find serializer for type $type");
     }
 
-    return serializer.serialize(object,
-        withTypeInfo: withTypeInfo, typeInfoKey: useTypeInfoKey);
+    return serializer.serialize(object, withType: withType, typeKey: typeKey);
   }
 
-  dynamic _from(dynamic decoded, {Type type, String useTypeInfoKey}) {
+  dynamic _from(dynamic decoded, {Type type, String typeKey}) {
     Serializer serializer;
 
     try {
       serializer = getByType(type);
     } catch (e) {
-      final String typeInfo = _objectToType(decoded, useTypeInfoKey);
+      final String typeInfo = _objectToType(decoded, typeKey);
       if (typeInfo is String) {
         serializer = getByTypeKey(typeInfo);
       }