+++ lib/src/yaml/yaml.dart
@@ -10,8 +10,7 @@ import 'package:yamlicious/yamlicious.dart';
  * Same usage as [SerializerRepo]
  */
 class YamlRepo extends SerializerRepo {
-  YamlRepo({String typeInfoKey: defaultTypeInfoKey})
-      : super(typeInfoKey: typeInfoKey);
+  YamlRepo({String typeKey: defaultTypeInfoKey}) : super(typeKey: typeKey);
 
   ///@nodoc
   dynamic encode(dynamic object) => toYamlString(object);
@@ -27,8 +26,8 @@ class YamlRepo extends SerializerRepo {
    * See [SerializerRepo.from] for more information.
    */
   @override
-  dynamic deserialize(dynamic object, {Type type, String useTypeInfoKey}) =>
-      super.deserialize(object, type: type, useTypeInfoKey: useTypeInfoKey);
+  dynamic deserialize(dynamic object, {Type type, String typeKey}) =>
+      super.deserialize(object, type: type, typeKey: typeKey);
 
   /**
    * Serialize an object to a YAML String
@@ -38,8 +37,6 @@ class YamlRepo extends SerializerRepo {
    * See [SerializerRepo.to] for more information.
    */
   @override
-  dynamic serialize(dynamic object,
-          {bool withTypeInfo: false, String useTypeInfoKey}) =>
-      super.serialize(object,
-          withTypeInfo: withTypeInfo, useTypeInfoKey: useTypeInfoKey);
+  dynamic serialize(dynamic object, {bool withType: false, String typeKey}) =>
+      super.serialize(object, withType: withType, typeKey: typeKey);
 }